Update user exp and fix some button and routes bugs

+++ client/src/actions/post.js
@@ -1,11 +1,17 @@
 import api from '../utils/api';
 import { setAlert } from './alert';
 import {
+  ADD_COMMENT,
+  REMOVE_COMMENT,
+  ADD_COMMENT_ARRAY,
+  REMOVE_COMMENT_ARRAY,
   ADD_POST,
   DELETE_POST,
   GET_POSTS,
+  GET_POST,
   POST_ERROR,
   UPDATE_LIKES,
+  SINGLE_POST_UPDATE_LIKES,
 } from './types';
 
 // Get all posts
@@ -27,14 +33,8 @@ export const getPosts = () => async (dispatch) => {
 
 // Add Post
 export const addPost = (formData) => async (dispatch) => {
-  const config = {
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  };
-
   try {
-    const res = await api.post('/posts', formData, config);
+    const res = await api.post('/posts', formData);
 
     dispatch({
       type: ADD_POST,
@@ -52,21 +52,23 @@ export const addPost = (formData) => async (dispatch) => {
 
 // Delete Post
 export const deletePost = (id) => async (dispatch) => {
-  try {
-    console.log('Post deleted');
-    await api.delete(`/posts/${id}`);
+  if (window.confirm('Are you sure to delete your post?')) {
+    try {
+      console.log('Post deleted');
+      await api.delete(`/posts/${id}`);
 
-    dispatch({
-      type: DELETE_POST,
-      payload: id,
-    });
+      dispatch({
+        type: DELETE_POST,
+        payload: id,
+      });
 
-    dispatch(setAlert('Post Deleted', 'success'));
-  } catch (err) {
-    dispatch({
-      type: POST_ERROR,
-      payload: { msg: err.response.statusText, status: err.response.status },
-    });
+      dispatch(setAlert('Post Deleted', 'success'));
+    } catch (err) {
+      dispatch({
+        type: POST_ERROR,
+        payload: { msg: err.response.statusText, status: err.response.status },
+      });
+    }
   }
 };
 
@@ -95,6 +97,136 @@ export const removeLike = (id) => async (dispatch) => {
     dispatch({
       type: UPDATE_LIKES,
       payload: { id, likes: res.data },
+    });
+  } catch (err) {
+    dispatch({
+      type: POST_ERROR,
+      payload: { msg: err.response.statusText, status: err.response.status },
+    });
+  }
+};
+
+// ## This is for SINGLE POST page
+// Single post add like
+export const singleAddLike = (id) => async (dispatch) => {
+  try {
+    const res = await api.put(`/posts/like/${id}`);
+
+    dispatch({
+      type: SINGLE_POST_UPDATE_LIKES,
+      payload: { id, likes: res.data },
+    });
+  } catch (err) {
+    dispatch({
+      type: POST_ERROR,
+      payload: { msg: err.response.statusText, status: err.response.status },
+    });
+  }
+};
+
+// Single post remove like
+export const singleRemoveLike = (id) => async (dispatch) => {
+  try {
+    const res = await api.put(`/posts/unlike/${id}`);
+
+    dispatch({
+      type: SINGLE_POST_UPDATE_LIKES,
+      payload: { id, likes: res.data },
+    });
+  } catch (err) {
+    dispatch({
+      type: POST_ERROR,
+      payload: { msg: err.response.statusText, status: err.response.status },
+    });
+  }
+};
+
+// Add Comment
+export const addComment = (postId, formData) => async (dispatch) => {
+  try {
+    const res = await api.post(`/posts/comment/${postId}`, formData);
+
+    dispatch({
+      type: ADD_COMMENT,
+      payload: res.data,
+    });
+
+    dispatch(setAlert('Comment Added', 'success'));
+  } catch (err) {
+    dispatch({
+      type: POST_ERROR,
+      payload: { msg: err.response.statusText, status: err.response.status },
+    });
+  }
+};
+
+// Delete comment
+export const removeComment = (postId, commentId) => async (dispatch) => {
+  try {
+    await api.delete(`/posts/comment/${postId}/${commentId}`);
+
+    dispatch({
+      type: REMOVE_COMMENT,
+      payload: commentId,
+    });
+
+    dispatch(setAlert('Comment Removed', 'success'));
+  } catch (err) {
+    dispatch({
+      type: POST_ERROR,
+      payload: { msg: err.response.statusText, status: err.response.status },
+    });
+  }
+};
+
+// Add Comment in Array
+export const addCommentArray = (postId, formData) => async (dispatch) => {
+  try {
+    const res = await api.post(`/posts/comment/${postId}`, formData);
+
+    dispatch({
+      type: ADD_COMMENT_ARRAY,
+      payload: { postId, comments: res.data },
+    });
+
+    console.log(postId, res.data);
+
+    dispatch(setAlert('Comment Added', 'success'));
+  } catch (err) {
+    dispatch({
+      type: POST_ERROR,
+      payload: { msg: err.response.statusText, status: err.response.status },
+    });
+  }
+};
+
+// Remove Comment in Array
+export const removeCommentArray = (postId, commentId) => async (dispatch) => {
+  try {
+    await api.delete(`/posts/comment/${postId}/${commentId}`);
+
+    dispatch({
+      type: REMOVE_COMMENT_ARRAY,
+      payload: commentId,
+    });
+
+    dispatch(setAlert('Comment Removed', 'success'));
+    // console.log(dispatch({ payload: commentId }));
+  } catch (err) {
+    dispatch({
+      type: POST_ERROR,
+      payload: { msg: err.response.statusText, status: err.response.status },
+    });
+  }
+};
+
+export const getPost = (id) => async (dispatch) => {
+  try {
+    const res = await api.get(`/posts/${id}`);
+
+    dispatch({
+      type: GET_POST,
+      payload: res.data,
     });
   } catch (err) {
     dispatch({
